+++ golem-worker-service-base/src/api/custom_http_request_api.rs
@@ -12,7 +12,7 @@ use crate::service::http_request_definition_lookup::ApiDefinitionLookup;
 
 use crate::worker_binding::worker_binding_resolver::WorkerBindingResolver;
 use crate::worker_bridge::WorkerRequest;
-use crate::worker_bridge::workr_request_executor::WorkerRequestExecutor;
+use crate::worker_bridge::worker_request_executor::WorkerRequestExecutor;
 
 // Executes custom request with the help of worker_request_executor and definition_service
 // This is a common API projects can make use of, similar to healthcheck service
